fix ts tests
